pyproject.toml
[project]
name = "zero-network"
version = "0.1.0"
description = "Python SDK for the Zero Network — stablecoin microtransactions for AI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{name = "Zero Network", email = "dev@zzero.net"}]
keywords = ["zero", "micropayments", "stablecoin", "ai-agents", "x402"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "httpx>=0.24.0",
    "pynacl>=1.5.0",
]

[project.optional-dependencies]
x402 = ["starlette>=0.20.0"]

[project.urls]
Homepage = "https://zzero.net"
Documentation = "https://docs.zzero.net"
Repository = "https://github.com/Zzero-net/zero-sdk-python"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
